[WS] rename ChatGateway into Gateway
