+++ src/nestjs/src/app.module.ts
@@ -5,7 +5,7 @@ import { APP_GUARD } from "@nestjs/core";
 import { DBModule } from "./modules/database/database.module";
 import { AuthModule } from "./modules/auth/module";
 import { JwtAuthGuard } from "./modules/auth/jwt-auth.guard";
-import { WSChatModule } from "./websocket/chat/chat.module";
+import { WSModule } from "./websocket/ws.module";
 import { UserModule } from "./adapter/user/module";
 import { ChatRoomModule } from "./adapter/chatRoom/module";
 
@@ -14,7 +14,7 @@ import { ChatRoomModule } from "./adapter/chatRoom/module";
 		ConfigModule.forRoot({ isGlobal: true }),
 		DBModule,
 		AuthModule,
-		WSChatModule,
+		WSModule,
 		UserModule,
 		ChatRoomModule,
 	],
