+++ src/nestjs/src/websocket/ws.gateway.ts
@@ -7,13 +7,13 @@ import {
 	WebSocketServer,
 } from "@nestjs/websockets";
 import { Server, Socket } from "socket.io";
-import { WSChatService } from "./chat.service";
+import { WSChatService } from "./chat/chat.service";
 
 @WebSocketGateway(3001, {
 	path: "/ws",
 	cors: { origin: "*" },
 })
-export class WSChatGateway
+export class WSGateway
 	implements OnGatewayConnection, OnGatewayDisconnect
 {
 	constructor(private wsChatService: WSChatService) {}
